02-07-23_Update

+++ src/map/lua/luautils.h
@@ -176,7 +176,7 @@ namespace luautils
     int32 terminate(lua_State*);                                                // Logs off all characters and terminates the server
 
     int32 GetTextIDVariable(uint16 ZoneID, const char* variable);               // загружаем значение переменной TextID указанной зоны
-    uint8 GetSettingsVariable(const char* variable);                            // Gets a Variable Value from Settings.lua
+    float GetSettingsVariable(const char* variable);                            // Gets a Variable Value from Settings.lua
     bool IsContentEnabled(const char* content);                                 // Check if the content is enabled in settings.lua
 
     int32 OnGameDay(CZone* PZone);                                              // Automatic action of NPC every game day
