+++ src/map/utils/zoneutils.cpp
@@ -342,8 +342,8 @@ void LoadNPCList()
 
 void LoadMOBList()
 {
-    uint8 normalLevelRangeMin = luautils::GetSettingsVariable("NORMAL_MOB_MAX_LEVEL_RANGE_MIN");
-    uint8 normalLevelRangeMax = luautils::GetSettingsVariable("NORMAL_MOB_MAX_LEVEL_RANGE_MAX");
+    uint8 normalLevelRangeMin = (uint8)(luautils::GetSettingsVariable("NORMAL_MOB_MAX_LEVEL_RANGE_MIN"));
+    uint8 normalLevelRangeMax = (uint8)(luautils::GetSettingsVariable("NORMAL_MOB_MAX_LEVEL_RANGE_MAX"));
 
     const char* Query =
         "SELECT mob_groups.zoneid, mobname, mobid, pos_rot, pos_x, pos_y, pos_z, \
